Rename Account.ts -> AccountState.ts
